requirements.txt
streamlit
pandas
scikit-learn
joblib
numpy
pillow
opencv-python-headless

